refactor: simplify Tabby profile synchronization
