+++ .agents/scripts/tabby-profile-sync.py
@@ -20,12 +20,25 @@ import json
 import os
 import re
 import subprocess
-import sys
 import uuid
 from pathlib import Path
-from typing import NamedTuple
 
 from tabby_colour_utils import generate_tab_colour, find_closest_scheme
+from tabby_profile_repair import (
+    _repair_broken_opencode_launch_profile_block,
+    repair_broken_opencode_launch_profiles,
+)
+from tabby_profile_utils import (
+    LEGACY_TABBY_COMMAND_FIELD_OPENCODE,
+    LEGACY_TABBY_OPENCODE_LAUNCH,
+    TABBY_COMMAND_FIELD_OPENCODE,
+    TABBY_OPENCODE_LAUNCH,
+)
+from tabby_profile_validation import (
+    ProfileArgTypeIssue,
+    find_profile_arg_type_issues,
+    report_profile_arg_type_issues,
+)
 from tabby_yaml_helpers import (
     load_yaml_simple,
     save_yaml,
@@ -33,20 +46,6 @@ from tabby_yaml_helpers import (
     extract_group_id,
     insert_profiles_block,
 )
-
-
-TABBY_OPENCODE_LAUNCH = "aidevops opencode; exec zsh"
-TABBY_COMMAND_FIELD_OPENCODE = f"/bin/zsh -l -c '{TABBY_OPENCODE_LAUNCH}'"
-LEGACY_TABBY_OPENCODE_LAUNCH = "opencode; exec zsh"
-LEGACY_TABBY_COMMAND_FIELD_OPENCODE = "/bin/zsh -l -c 'opencode; exec zsh'"
-
-
-class ProfileArgTypeIssue(NamedTuple):
-    """A non-string value persisted in a Tabby profile's ``options.args``."""
-
-    profile_name: str
-    line_number: int
-    value: str
 
 
 def is_linked_worktree(repo_path: str) -> bool:
@@ -162,512 +161,6 @@ def build_group_yaml(group_id: str) -> str:
     """Build a YAML group block."""
     return f"""  - id: {group_id}
     name: Projects"""
-
-
-def _normalise_yaml_scalar(value: str) -> str:
-    """Return a plain scalar value from a simple YAML string/list item."""
-    in_single_quote = False
-    in_double_quote = False
-    comment_start = len(value)
-    for index, char in enumerate(value):
-        if char == "'" and not in_double_quote:
-            in_single_quote = not in_single_quote
-        elif char == '"' and not in_single_quote:
-            in_double_quote = not in_double_quote
-        elif char == "#" and not in_single_quote and not in_double_quote:
-            comment_start = index
-            break
-    value = value[:comment_start].strip()
-    if len(value) >= 2 and value[0] == value[-1] and value[0] in ("'", '"'):
-        return value[1:-1]
-    return value
-
-
-def _next_option_line(lines: list[str], start: int) -> str:
-    """Return the next non-comment line from ``start`` or an empty string."""
-    while start < len(lines):
-        line = lines[start]
-        if line.strip() and not line.lstrip().startswith("#"):
-            return line
-        start += 1
-    return ""
-
-
-def _parse_inline_args(value: str) -> list[str] | None:
-    """Parse a simple inline YAML list such as ``['-l', '-i']``."""
-    value = value.strip()
-    if not (value.startswith("[") and value.endswith("]")):
-        return None
-    inner = value[1:-1].strip()
-    if not inner:
-        return []
-    return [_normalise_yaml_scalar(part) for part in inner.split(",")]
-
-
-def _parse_block_args(lines: list[str]) -> list[str]:
-    """Parse simple ``- value`` YAML list entries from an args block."""
-    args: list[str] = []
-    for line in lines:
-        match = re.match(r"^\s*-\s*(.+?)\s*$", line)
-        if match:
-            args.append(_normalise_yaml_scalar(match.group(1)))
-    return args
-
-
-def _is_broken_opencode_args(args: list[str]) -> bool:
-    """Return True for the Tabby launch shape that breaks zsh job control."""
-    return args == ["-l", "-i", "-c", "opencode"]
-
-
-def _is_command_field_opencode(value: str) -> bool:
-    """Return True for the Tabby command-field shape that Tabby cannot exec."""
-    value = value.strip()
-    normalised = _normalise_yaml_scalar(value)
-    return (
-        value == TABBY_COMMAND_FIELD_OPENCODE
-        or value == LEGACY_TABBY_COMMAND_FIELD_OPENCODE
-        or normalised == TABBY_COMMAND_FIELD_OPENCODE
-        or normalised == LEGACY_TABBY_COMMAND_FIELD_OPENCODE
-    )
-
-
-def _is_legacy_direct_opencode_args(args: list[str]) -> bool:
-    """Return True for direct OpenCode args that still use the shared DB."""
-    return args == ["-l", "-c", LEGACY_TABBY_OPENCODE_LAUNCH]
-
-
-def _direct_opencode_args_block(args_indent: str, include_env: bool) -> list[str]:
-    """Build the direct Tabby args/env block for OpenCode profiles."""
-    child_indent = f"{args_indent}  "
-    block = [
-        f"{args_indent}args:",
-        f"{child_indent}- '-l'",
-        f"{child_indent}- '-c'",
-        f"{child_indent}- '{TABBY_OPENCODE_LAUNCH}'",
-    ]
-    if include_env:
-        block.append(f"{args_indent}env: {{}}")
-    return block
-
-
-def _line_indent_len(line: str) -> int:
-    """Return the number of leading spaces in ``line``."""
-    return len(line) - len(line.lstrip(" "))
-
-
-def _option_key_match(line: str) -> tuple[str, str] | None:
-    """Return ``(indent, key)`` for option-level keys we manage."""
-    match = re.match(r"^(?P<indent>\s*)(?P<key>args|command|env):(?:\s|$)", line)
-    if not match:
-        return None
-    return match.group("indent"), match.group("key")
-
-
-def _line_is_option_key(line: str, indent: str, key: str) -> bool:
-    """Return True if ``line`` is ``key`` at ``indent``."""
-    return bool(re.match(rf"^{re.escape(indent)}{re.escape(key)}:(?:\s|$)", line))
-
-
-def _has_prior_option_key(repaired: list[str], indent: str, key: str) -> bool:
-    """Return True if ``key`` already appeared in the current options block."""
-    indent_len = len(indent)
-    for previous in reversed(repaired):
-        if previous.strip() == "options:":
-            return False
-        if previous.strip() and _line_indent_len(previous) < indent_len:
-            return False
-        if _line_is_option_key(previous, indent, key):
-            return True
-    return False
-
-
-def _block_end(lines: list[str], start: int, base_indent_len: int) -> int:
-    """Return the first line after the scalar/list/mapping block at ``start``."""
-    block_end = start + 1
-    while block_end < len(lines):
-        next_line = lines[block_end]
-        if next_line.strip() and _line_indent_len(next_line) <= base_indent_len:
-            break
-        block_end += 1
-    return block_end
-
-
-def _tabby_autorun_env_end(lines: list[str], start: int, base_indent_len: int) -> int | None:
-    """Return env block end if it only carries ``TABBY_AUTORUN=opencode``."""
-    if start >= len(lines):
-        return None
-    env_line = lines[start]
-    env_indent_len = len(env_line) - len(env_line.lstrip(" "))
-    if env_indent_len != base_indent_len or env_line.strip() != "env:":
-        return None
-
-    block_end = start + 1
-    has_autorun = False
-    has_other_env = False
-    while block_end < len(lines):
-        next_line = lines[block_end]
-        if next_line.strip():
-            next_indent_len = len(next_line) - len(next_line.lstrip(" "))
-            if next_indent_len <= env_indent_len:
-                break
-            stripped = next_line.strip().strip("'").strip('"')
-            if stripped == "TABBY_AUTORUN: opencode":
-                has_autorun = True
-            else:
-                has_other_env = True
-        block_end += 1
-
-    if has_autorun and not has_other_env:
-        return block_end
-    return None
-
-
-def _line_mentions_opencode_launch(line: str) -> bool:
-    """Return True when a non-comment YAML line mentions an OpenCode launch."""
-    stripped = line.strip()
-    return bool(stripped and not stripped.startswith("#") and "opencode" in stripped)
-
-
-def _profile_mentions_opencode_launch(profile_lines: list[str]) -> bool:
-    """Return True when a profile block clearly targets OpenCode."""
-    return any(_line_mentions_opencode_launch(line) for line in profile_lines)
-
-
-def _profile_block_end(lines: list[str], start: int) -> int:
-    """Return the first line after the Tabby profile block at ``start``."""
-    return _block_end(lines, start, 2)
-
-
-def _is_non_string_yaml_list_item(value: str) -> bool:
-    """Return True when a simple YAML list item is not a string scalar."""
-    value = value.strip()
-    if not value or value.startswith(("{", "[")):
-        return True
-    if re.fullmatch(r"'(?:[^']|'')*'", value) or re.fullmatch(
-        r'"(?:[^"\\]|\\.)*"', value
-    ):
-        return False
-    if re.search(r":(?:\s|$)", value):
-        return True
-    if value.lower() in {"null", "true", "false", "~"}:
-        return True
-    return bool(re.fullmatch(r"[-+]?(?:\d[\d_]*)(?:\.\d[\d_]*)?", value))
-
-
-def _find_block_arg_type_issues(
-    lines: list[str],
-    start: int,
-    end: int,
-    expected_indent: int,
-    profile_name: str,
-) -> list[ProfileArgTypeIssue]:
-    """Find non-string list entries in one block-style ``args`` value."""
-    issues: list[ProfileArgTypeIssue] = []
-    for line_index in range(start, end):
-        line = lines[line_index]
-        if _line_indent_len(line) != expected_indent:
-            continue
-        item_match = re.match(r"^\s*-\s*(?P<value>.*)$", line)
-        if not item_match:
-            continue
-        value = item_match.group("value").strip()
-        if not _is_non_string_yaml_list_item(value):
-            continue
-        issues.append(ProfileArgTypeIssue(profile_name, line_index + 1, value))
-    return issues
-
-
-def _find_profile_arg_type_issues(
-    lines: list[str], profile_start: int, profile_end: int, profile_name: str
-) -> list[ProfileArgTypeIssue]:
-    """Find invalid ``args`` entries inside one Tabby profile block."""
-    issues: list[ProfileArgTypeIssue] = []
-    index = profile_start + 1
-    while index < profile_end:
-        args_match = re.match(r"^(?P<indent>\s*)args:\s*(?P<value>.*)$", lines[index])
-        if not args_match:
-            index += 1
-            continue
-
-        inline_value = args_match.group("value").strip()
-        if inline_value and "{" in inline_value:
-            issues.append(ProfileArgTypeIssue(profile_name, index + 1, inline_value))
-            index += 1
-            continue
-
-        args_indent_len = len(args_match.group("indent"))
-        args_end = _block_end(lines, index, args_indent_len)
-        issues.extend(
-            _find_block_arg_type_issues(
-                lines,
-                index + 1,
-                args_end,
-                args_indent_len + 2,
-                profile_name,
-            )
-        )
-        index = args_end
-    return issues
-
-
-def find_profile_arg_type_issues(config_text: str) -> list[ProfileArgTypeIssue]:
-    """Find profile args that Tabby cannot pass to its PTY as ``string[]``.
-
-    Tabby's command-line editor uses ``shell-quote``. Pasting an unwrapped shell
-    pipeline makes the parser persist control operators as mappings such as
-    ``- op: '&&'``. The local PTY contract only accepts strings, so attempting
-    to open that profile can leave its renderer unresponsive.
-    """
-    lines = config_text.split("\n")
-    issues: list[ProfileArgTypeIssue] = []
-    i = 0
-    while i < len(lines):
-        profile_match = re.match(r"^  - name:\s*(?P<name>.+?)\s*$", lines[i])
-        if not profile_match:
-            i += 1
-            continue
-
-        profile_name = _normalise_yaml_scalar(profile_match.group("name"))
-        profile_end = _profile_block_end(lines, i)
-        issues.extend(
-            _find_profile_arg_type_issues(lines, i, profile_end, profile_name)
-        )
-        i = profile_end
-    return issues
-
-
-def report_profile_arg_type_issues(config_text: str) -> bool:
-    """Print actionable profile-schema errors and return whether any exist."""
-    issues = find_profile_arg_type_issues(config_text)
-    if not issues:
-        return False
-
-    locations: dict[str, list[int]] = {}
-    for issue in issues:
-        locations.setdefault(issue.profile_name, []).append(issue.line_number)
-
-    print(
-        "Invalid Tabby profile configuration: options.args must contain only strings.",
-        file=sys.stderr,
-    )
-    for profile_name, line_numbers in locations.items():
-        lines_text = ", ".join(str(line_number) for line_number in line_numbers)
-        print(
-            f"  - {profile_name}: non-string argument at line(s) {lines_text}",
-            file=sys.stderr,
-        )
-    print(
-        "Do not launch the listed profiles. In Tabby's command-line field, paste "
-        "the full quoted /bin/zsh -l -c '<command>' invocation so shell operators "
-        "remain inside one string argument.",
-        file=sys.stderr,
-    )
-    return True
-
-
-def _enable_dynamic_title(profile_text: str) -> tuple[str, bool]:
-    """Allow OSC title updates for one aidevops-managed OpenCode profile."""
-    lines = profile_text.split("\n")
-    for index, line in enumerate(lines):
-        if re.match(r"^    disableDynamicTitle:\s*", line):
-            replacement = "    disableDynamicTitle: false"
-            if line == replacement:
-                return profile_text, False
-            lines[index] = replacement
-            return "\n".join(lines), True
-
-    insert_at = len(lines)
-    for index, line in enumerate(lines):
-        if re.match(r"^    type:\s*", line):
-            insert_at = index
-            break
-    lines.insert(insert_at, "    disableDynamicTitle: false")
-    return "\n".join(lines), True
-
-
-def _repair_broken_opencode_launch_profile_block(config_text: str) -> tuple[str, int]:
-    """Repair fragile Tabby OpenCode launch profiles inside one profile block.
-
-    ``zsh -i -c`` enables interactive startup while executing a command string,
-    which can trigger Powerlevel10k/gitstatus job-control errors before the TUI
-    starts. The former ``TABBY_AUTORUN`` workaround can also fail silently when
-    shell startup does not run the hook, leaving users in a plain terminal. The
-    stable shape uses a non-interactive login command and leaves zsh open after
-    OpenCode exits.
-    """
-    lines = config_text.split("\n")
-    repaired: list[str] = []
-    repairs = 0
-    pending_command_field_indent: str | None = None
-    i = 0
-
-    while i < len(lines):
-        line = lines[i]
-
-        command_match = re.match(r"^(?P<indent>\s*)command:\s*(?P<value>.*)$", line)
-        if command_match and _is_command_field_opencode(command_match.group("value")):
-            command_indent = command_match.group("indent")
-            if _has_prior_option_key(repaired, command_indent, "command"):
-                i = _block_end(lines, i, len(command_indent))
-                repairs += 1
-                continue
-            repaired.append(f"{command_indent}command: /bin/zsh")
-            pending_command_field_indent = command_indent
-            repairs += 1
-            i += 1
-            continue
-
-        option_key = _option_key_match(line)
-        if option_key is not None:
-            option_indent, key = option_key
-            if key != "args" or pending_command_field_indent != option_indent:
-                if _has_prior_option_key(repaired, option_indent, key):
-                    i = _block_end(lines, i, len(option_indent))
-                    repairs += 1
-                    continue
-
-        if pending_command_field_indent is not None:
-            pending_indent_len = len(pending_command_field_indent)
-            if (
-                line.strip()
-                and _line_indent_len(line) <= pending_indent_len
-                and not _line_is_option_key(line, pending_command_field_indent, "args")
-            ):
-                include_env = not _line_is_option_key(
-                    line, pending_command_field_indent, "env"
-                ) and not _has_prior_option_key(
-                    repaired, pending_command_field_indent, "env"
-                )
-                repaired.extend(
-                    _direct_opencode_args_block(
-                        pending_command_field_indent,
-                        include_env=include_env,
-                    )
-                )
-                pending_command_field_indent = None
-
-        match = re.match(r"^(?P<indent>\s*)args:\s*(?P<value>.*)$", line)
-        if not match:
-            repaired.append(line)
-            i += 1
-            continue
-
-        args_indent = match.group("indent")
-        args_indent_len = len(args_indent)
-        inline_args = _parse_inline_args(match.group("value"))
-        if inline_args is not None:
-            if pending_command_field_indent == args_indent:
-                next_line = _next_option_line(lines, i + 1)
-                include_env = not _line_is_option_key(
-                    next_line, args_indent, "env"
-                ) and not _has_prior_option_key(repaired, args_indent, "env")
-                repaired.extend(
-                    _direct_opencode_args_block(args_indent, include_env=include_env)
-                )
-                pending_command_field_indent = None
-            elif _is_broken_opencode_args(inline_args) or _is_legacy_direct_opencode_args(inline_args):
-                next_line = _next_option_line(lines, i + 1)
-                include_env = not _line_is_option_key(
-                    next_line, args_indent, "env"
-                ) and not _has_prior_option_key(repaired, args_indent, "env")
-                repaired.extend(
-                    _direct_opencode_args_block(args_indent, include_env=include_env)
-                )
-                repairs += 1
-            elif inline_args == ["-l", "-i"]:
-                env_end = _tabby_autorun_env_end(lines, i + 1, args_indent_len)
-                if env_end is not None:
-                    include_env = not _has_prior_option_key(repaired, args_indent, "env")
-                    repaired.extend(
-                        _direct_opencode_args_block(args_indent, include_env=include_env)
-                    )
-                    repairs += 1
-                    i = env_end
-                    continue
-                repaired.append(line)
-            else:
-                repaired.append(line)
-            i += 1
-            continue
-
-        block_end = _block_end(lines, i, args_indent_len)
-
-        block_args = _parse_block_args(lines[i + 1 : block_end])
-        if pending_command_field_indent == args_indent:
-            next_line = _next_option_line(lines, block_end)
-            include_env = not _line_is_option_key(
-                next_line, args_indent, "env"
-            ) and not _has_prior_option_key(repaired, args_indent, "env")
-            repaired.extend(_direct_opencode_args_block(args_indent, include_env=include_env))
-            pending_command_field_indent = None
-        elif _is_broken_opencode_args(block_args) or _is_legacy_direct_opencode_args(block_args):
-            next_line = _next_option_line(lines, block_end)
-            include_env = not _line_is_option_key(
-                next_line, args_indent, "env"
-            ) and not _has_prior_option_key(repaired, args_indent, "env")
-            repaired.extend(_direct_opencode_args_block(args_indent, include_env=include_env))
-            repairs += 1
-        elif block_args == ["-l", "-i"]:
-            env_end = _tabby_autorun_env_end(lines, block_end, args_indent_len)
-            if env_end is not None:
-                include_env = not _has_prior_option_key(repaired, args_indent, "env")
-                repaired.extend(_direct_opencode_args_block(args_indent, include_env=include_env))
-                repairs += 1
-                i = env_end
-                continue
-            repaired.extend(lines[i:block_end])
-        else:
-            repaired.extend(lines[i:block_end])
-        i = block_end
-
-    if pending_command_field_indent is not None:
-        include_env = not _has_prior_option_key(repaired, pending_command_field_indent, "env")
-        repaired.extend(
-            _direct_opencode_args_block(
-                pending_command_field_indent,
-                include_env=include_env,
-            )
-        )
-
-    return "\n".join(repaired), repairs
-
-
-def repair_broken_opencode_launch_profiles(config_text: str) -> tuple[str, int]:
-    """Repair fragile OpenCode profiles without touching custom profiles.
-
-    Tabby configs can contain hand-authored profiles for local web servers,
-    shells, and other commands. The OpenCode repair logic manages ``command``,
-    ``args``, and ``env`` keys, so scope it to profile blocks that explicitly
-    mention OpenCode and preserve all other profile blocks byte-for-byte.
-    """
-    lines = config_text.split("\n")
-    repaired: list[str] = []
-    repairs = 0
-    i = 0
-
-    while i < len(lines):
-        line = lines[i]
-        if not line.startswith("  - name:"):
-            repaired.append(line)
-            i += 1
-            continue
-
-        block_end = _profile_block_end(lines, i)
-        profile_lines = lines[i:block_end]
-        if not _profile_mentions_opencode_launch(profile_lines):
-            repaired.extend(profile_lines)
-            i = block_end
-            continue
-
-        original_block = "\n".join(profile_lines)
-        repaired_block, repaired_count = _repair_broken_opencode_launch_profile_block(original_block)
-        repaired_block, dynamic_title_changed = _enable_dynamic_title(repaired_block)
-        repaired.extend(repaired_block.split("\n"))
-        if repaired_count > 0 or dynamic_title_changed:
-            repairs += 1
-        i = block_end
-
-    return "\n".join(repaired), repairs
 
 
 def ensure_groups_section(config_text: str, group_id: str) -> str:
